Bug fixes and results input search
The input search is working but i have an issue with re-renders and states from Redux. On every input change, results are updated automatically with new results, even when searchSubmit() is not called. I don't know why is this behavior happening.
On the other hand, when i will be able to fix this issue, i would have to save country and region states for betters searches.